tbl died on 16 of its authors' 54 tables, and wavec was green
Task #91.  The three author-written suites unlocked by 30169a9 are wired
into tests/wavec: eqn 35 cases, pic 36, tbl 54.  eqn and pic were clean
from the first run.  tbl was not.

SIXTEEN OF FIFTY-FOUR DIED ON SIGSEGV -- 30% -- while this suite had been
green for the life of the port.  Both causes are the same one-word defect,
an int holding a pointer, and the crash addresses give the class away:
0x1e6058a9 and 0x4ac8141, both plainly 32-bit values.

  tm.c    `int dpoint' carries the ADDRESS of a decimal point in maknew(),
          and is also tested as a boolean four times -- upstream's
          flag-and-pointer idiom, exact on a VAX.  IT IS THE `n' COLUMN,
          tbl's characteristic feature, and maknew is called for nothing
          else (t5.c:72, t9.c:49).  Twenty bytes reproduce it.  l, c, r and
          a columns are all clean, measured, which is exactly why it
          survived: the hand-written tables here had never used an `n'.

  t0/t5/t..c  `int leftover' carries the address of the line that did not
          fit.  t9.c:14 passes it to `domore(dataln) char *dataln;' which
          hands it to prefix(), which dereferences it.  THREE declarations
          move, not one -- the definition, the extern in t..c (an
          #include'd non-header, so every object sees it), and the cast.

AND THE CAST HAS TO MOVE WITH THE DECLARATION.  `long dpoint' alone fixes
nothing, because `(int)str' truncates BEFORE the value is widened into the
long: the explicit cast is the truncation, not the storage.  Same trap in
both files.

`long' rather than `char *' in both, because the boolean uses are
upstream's and a pointer type needs three more edits each to keep them --
the same one-word fix as yacc's YYSTYPE and qed's signal handlers.

AUDITED AND NOT CHANGED: t8.c:302 is the identical idiom and needs no fix.
point() at tc.c:57 is `s >= 128 || s < 0', a pointer-or-marker test, and
every later use of s is %c on the MARKER branch -- the pointer branch
continues.  A truncated pointer changes no answer unless its low 32 bits
fall under 128.  make/meter() precedent.

STILL BROKEN, DELIBERATELY NAMED: sample32, a DIFFERENT class -- permute()
at t5.c:124 writes through a null row pointer (0x8, not a 32-bit value)
walking a `^' vertical span in a table that also has full-width rules.
wavec asserts it STILL dies, so fixing it means removing the name from
KNOWN_TBL_FAIL in the same commit, which is the crash probe's discipline.

NO ORACLE, DELIBERATELY.  Bell Labs' own harness (the `compare' member of
eqntest.a) diffs a NEW eqn against the INSTALLED one -- differential
against the previous version, which this port has not got.  Freezing
today's output as golden would rebuild the trap this suite has already
fallen into twice.  What is asserted is what can be asserted honestly: no
signal death, non-empty output, determinism, and the n column working.

Mutation: reverting both fixes fires exactly two cases and names all 15
regressed samples.

wavec 82 -> 89; make test 2422 -> 2429, 0 failed.

+++ v8/src/cmd/tbl/t..c
@@ -51,7 +51,21 @@ extern char *exstore, *exlim, *exspace;
 extern int *sep;
 extern int *used, *lused, *rused;
 extern int linestop[];
-extern int leftover;
+/*
+ * PORT: `long leftover', was `int'.  The tm.c defect a second time, in the
+ * same program and with the same flag-and-pointer idiom -- t5.c:25 is
+ * `leftover=(int)cstore' and t5.c:15 is `leftover=0', so one variable is both
+ * a boolean (t7.c:18 `if (leftover)') and the address of the line that did not
+ * fit.  t9.c:14 then passes it to `domore(dataln) char *dataln;' which hands
+ * it to prefix(), and prefix dereferences it.  Measured: EXC_BAD_ACCESS at
+ * 0x4ac8141, a 32-bit value, in prefix.
+ *
+ * BOTH DECLARATIONS MOVE, and this one is in an #include'd non-header, so
+ * every object in tbl sees it -- which is what makes the coupling real here
+ * where qed's savint had none.  The cast at t5.c:25 has to move too: an
+ * explicit (int) truncates before the value is widened into the long.
+ */
+extern long leftover;
 extern char *last, *ifile;
 extern int texname;
 extern int texct, texmax;
